Add files via upload
This project is all about generating google forms as a quiz. We have to give Question bank in excel and test details and also students and staff details. This project will create random question in a google form and it will send the test link to the respective student and staff using google script.

Done by uipath packages, google script.